Add image and group object models
