+++ src/models/group.model.js
@@ -1,0 +1,9 @@
+const groupObj = {
+  id: '',
+  thumbnailUrl: '',
+  imageUrl: '',
+  title: '',
+  description: '',
+};
+
+export default groupObj;
