+++ src/models/image.model.js
@@ -1,0 +1,10 @@
+const imageObj = {
+  id: '',
+  thumbnailUrl: '',
+  imageUrl: '',
+  title: '',
+  description: '',
+  location: '',
+};
+
+export default imageObj;
